refactor: add custom hook to reuse for order line locations
