+++ src/common/hooks/useOrderLineLocations/index.js
@@ -1,0 +1,1 @@
+export { useOrderLineLocations } from './useOrderLineLocations';
