+++ src/common/hooks/useOrderLineLocations/useOrderLineLocations.js
@@ -1,0 +1,52 @@
+import { useMemo } from 'react';
+
+import {
+  useCentralOrderingContext,
+  useHoldingsAndLocations,
+  useLocationsQuery,
+  useReceivingTenantIdsAndLocations,
+} from '@folio/stripes-acq-components';
+import { useStripes } from '@folio/stripes/core';
+
+export const useOrderLineLocations = ({ poLineLocations = [], instanceId }) => {
+  const stripes = useStripes();
+  const currentTenantId = stripes?.okapi?.tenant;
+
+  const { isCentralOrderingEnabled } = useCentralOrderingContext();
+
+  const {
+    isLoading: isLocationsLoading,
+    locations,
+  } = useLocationsQuery({ consortium: isCentralOrderingEnabled });
+
+  const receivingTenants = useMemo(() => {
+    if (poLineLocations.length) {
+      return poLineLocations.map(({ tenantId }) => tenantId);
+    }
+
+    return [];
+  }, [poLineLocations]);
+
+  const { receivingTenantIds } = useReceivingTenantIdsAndLocations({
+    receivingTenantIds: receivingTenants,
+    currentReceivingTenantId: currentTenantId,
+  });
+
+  const {
+    holdings,
+    locations: holdingLocations,
+    isLoading: isHoldingLocationsLoading,
+  } = useHoldingsAndLocations({
+    instanceId,
+    receivingTenantIds,
+    tenantId: currentTenantId,
+  });
+
+  const _poLineLocations = instanceId ? holdingLocations : locations;
+
+  return {
+    holdings,
+    locations: _poLineLocations,
+    isLoading: isLocationsLoading || isHoldingLocationsLoading,
+  };
+};
